% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{auto_grid}
\alias{auto_grid}
\title{Auto-generate a grid template with exactly
splitted in full numbers}
\usage{
auto_grid(x, grid_size = 100, chunks = 30L)
}
\arguments{
\item{x}{sf/terra object representing the polygon to be gridded}

\item{grid_size}{numeric(1). Size of the grid cells in meters.}

\item{chunks}{integer(1). Number of chunks to split the grid into.}
}
\value{
A SpatVector object with exactly split grids
}
\description{
Auto-generate a grid template with exactly
splitted in full numbers
}
