% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{create_inla_mesh}
\alias{create_inla_mesh}
\title{Create an INLA Mesh from sf Data}
\usage{
create_inla_mesh(data, formula, mesh, family = "gaussian", control = list())
}
\arguments{
\item{sf_data}{An \code{sf} object containing the spatial features to be meshed.}

\item{max_edge}{Numeric vector specifying the maximum allowed triangle
edge lengths in the mesh.}

\item{crs}{Coordinate reference system to use for the mesh (default: 4326).}
}
\value{
An INLA mesh object created using \code{fmesher::fm_mesh_2d_inla}.
}
\description{
This function generates a 2D mesh suitable for spatial modeling with INLA,
based on input spatial features (\code{sf} object).
}
\details{
The function transforms the input \code{sf} data to the specified CRS,
extracts coordinates, and defines a bounding box slightly larger than
the data extent. It then constructs a mesh using
\code{fmesher::fm_mesh_2d_inla}, with boundaries and interior segments derived
from the input features.
}
\examples{
\dontrun{
library(sf)
nc <- st_read(system.file("shape/nc.shp", package="sf"))
mesh <- create_inla_mesh(nc, max_edge = c(0.1, 0.5))
}
}
