% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{detrend}
\alias{detrend}
\title{Detrend using STL model}
\usage{
detrend(
  data,
  id,
  season_length = 8760,
  vars = NULL,
  result = "PM10",
  fill = FALSE
)
}
\arguments{
\item{data}{data.table sorted by timestamps.}

\item{id}{character(1). Site identifier.}

\item{season_length}{integer(1). Length of seasonal trend detection.}

\item{vars}{character.}

\item{result}{character(1). Result variable name.}

\item{fill}{logical(1).
Filling (imputing) missing values that present in the data.}
}
\value{
STL fit results.
}
\description{
Detrend using STL model
}
