% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extend_grid}
\alias{extend_grid}
\title{Extend space-time data.frame into full grid using start and end dates}
\usage{
extend_grid(data)
}
\arguments{
\item{data}{data.frame with columns TMSID2, TMSID, and date_s}
}
\value{
data.frame with complete grid of TMSID2, TMSID, and date_s
}
\description{
It takes a data frame with columns TMSID2, TMSID, date_s
then for each unique TMSID2-TMSID pair, it generates a complete
sequence of dates from date_start and date_end values.
This ensures that the resulting data frame
has a complete and unique grid of TMSID2, TMSID, and date_s,
which is essential for calculation while avoiding redundant rows.
}
