% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_all_tmb_spt}
\alias{fit_all_tmb_spt}
\title{Fit Spatial-Temporal Models and Predict Relocation Impact}
\usage{
fit_all_tmb_spt(
  sf_correct = sites_sf,
  sf_original = sites_asis_sf,
  formula = PM10 ~ 1,
  time = "year",
  year_target = 2020,
  resolution = 1000,
  reference = kor_all,
  likelihood = gaussian(),
  anisotropy = TRUE,
  log_transform = FALSE
)
}
\arguments{
\item{sf_correct}{A \code{sf} object containing corrected site data with spatial and PM10 information.}

\item{sf_original}{A \code{sf} object containing original site data with spatial and PM10 information.}

\item{formula}{A formula specifying the model to be fitted. Default is \code{PM10 ~ 1}.}

\item{time}{A string specifying the time variable in the dataset. Default is \code{"year"}.}

\item{year_target}{An integer specifying the target year for filtering data. Default is \code{2020}.}

\item{resolution}{A numeric value specifying the resolution of the prediction grid in meters. Default is \code{1000}.}

\item{reference}{A \code{sf} object used as the spatial reference for creating the prediction grid.}

\item{likelihood}{A family object specifying the likelihood function for the model. Default is \code{gaussian()}.}

\item{anisotropy}{A logical value indicating whether to include anisotropy in the model. Default is \code{TRUE}.}

\item{log_transform}{A logical value indicating whether to log-transform the predictions. Default is \code{FALSE}.}
}
\value{
A list containing the following elements:
\describe{
\item{pred_correct}{A raster object of predictions for the corrected site data.}
\item{pred_original}{A raster object of predictions for the original site data.}
\item{pred_diff}{A raster object of the difference between corrected and original predictions.}
}
}
\description{
This function fits spatial-temporal models using the \code{sdmTMB} package for two datasets:
corrected site data (\code{sf_correct}) and original site data (\code{sf_original}). It then predicts
the impact of relocation on a grid and computes the difference between the predictions.
}
\details{
The function performs the following steps:
\enumerate{
\item Creates a prediction grid based on the reference spatial object.
\item Prepares the corrected and original site data by filtering and adding coordinates.
\item Configures mesh objects for spatial modeling.
\item Fits spatial-temporal models for both corrected and original datasets.
\item Predicts PM10 values on the grid for both datasets.
\item Computes the difference between the predictions.
}
}
\examples{
\dontrun{
result <- fit_all_tmb_spt(
  sf_correct = sites_sf,
  sf_original = sites_asis_sf,
  formula = PM10 ~ 1,
  time = "year",
  year_target = 2020,
  resolution = 1000,
  reference = kor_all,
  likelihood = gaussian(),
  anisotropy = TRUE,
  log_transform = FALSE
)
}

}
