% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_tmb}
\alias{fit_tmb}
\title{Predict using a fitted TMB model on a spatial grid}
\usage{
fit_tmb(
  data,
  formula,
  formula_s = NULL,
  mesh,
  family,
  spatial = "on",
  reml = TRUE,
  anisotropy = TRUE,
  bayesian = TRUE,
  ...
)
}
\arguments{
\item{fit}{A fitted TMB model object.}

\item{grid}{An \code{sf} spatial object representing the prediction grid.}

\item{invars}{A character vector of variable names to include
in the prediction data frame.}
}
\value{
A vector or data frame of predictions from the fitted model.
}
\description{
This function generates predictions from a fitted TMB model (\code{fit}) over
a provided spatial grid (\code{grid}). It extracts spatial coordinates,
processes them, and constructs a new data frame with the required input
variables (\code{invars}) before making predictions.
}
\examples{
\dontrun{
predictions <- pred_tmb(fit, grid, c("var1", "var2"))
}
}
