% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{impute_df}
\alias{impute_df}
\title{Impute any missing values}
\usage{
impute_df(data)
}
\arguments{
\item{data}{data.frame. Data to be imputed.}
}
\value{
Imputed data.frame.
}
\description{
Impute any missing values
}
