% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{lower_filter}
\alias{lower_filter}
\title{Geographical smoothing of a raster}
\usage{
lower_filter(x, mat = NULL, rad)
}
\arguments{
\item{x}{A \code{SpatRaster} object from the \code{terra} package.
The raster to smooth.}

\item{mat}{A matrix defining the smoothing kernel.
If \code{NULL}, a default kernel is used.}
}
\value{
A \code{SpatRaster} object with smoothed values.
}
\description{
Smoothing filter for a raster object. It may be useful
for calculating relative altitude or other derivations
}
\details{
The function applies a focal operation to smooth the raster values
using the specified kernel. If no kernel is provided, a default
kernel is used.
}
\examples{
library(terra)
r <- rast(matrix(rnorm(100), 10, 10))
mat <- matrix(1, 3, 3)
smoothed_r <- detect_lower(r, mat)

}
