% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{pred_tmb}
\alias{pred_tmb}
\title{Predict using a fitted TMB model on a spatial grid}
\usage{
pred_tmb(fit, grid, invars)
}
\arguments{
\item{fit}{A fitted TMB model object.}

\item{grid}{An \code{sf} spatial object representing the prediction grid.}

\item{invars}{A character vector of variable names to include
in the prediction data frame.}
}
\value{
A vector or data frame of predictions from the fitted model.
}
\description{
This function generates predictions from a fitted TMB model (\code{fit}) over
a provided spatial grid (\code{grid}). It extracts spatial coordinates,
processes them, and constructs a new data frame with the required input
variables (\code{invars}) before making predictions.
}
\details{
The function extracts coordinates from the grid, constructs a data frame
with longitude, latitude, and the specified input variables (\code{invars}),
and then uses the \code{predict} function to generate predictions from
the fitted model.
}
\examples{
\dontrun{
predictions <- pred_tmb(fit, grid, c("var1", "var2"))
}
}
