% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{rasterize_freq_whitebox}
\alias{rasterize_freq_whitebox}
\title{Whitebox implementation of frequency rasterization}
\usage{
rasterize_freq_whitebox(ras, window_size = 67)
}
\arguments{
\item{ras}{A \code{SpatRaster} object from the \code{terra} package.}

\item{window_size}{Size of the moving window (must be odd).}
}
\value{
A \code{SpatRaster} object with frequency rasters for each class.
}
\description{
Whitebox implementation of frequency rasterization
}
\details{
This function uses the WhiteboxTools library to perform
a mean filter on binary masks for each class in the input raster.
}
