% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rectify_year}
\alias{rectify_year}
\title{Rectify year based on a date field}
\usage{
rectify_year(data, fieldname)
}
\arguments{
\item{data}{A data frame containing the date field.}

\item{fieldname}{The name of the date field in the data frame.}
}
\value{
A data frame with two new columns: \code{year_end} and \code{month_end}, and a rectified year column \code{year_rect}.
}
\description{
This function takes a data frame and a date field name, and rectifies the year based on the month of the date.
If the month is less than 6, it keeps the year as is; otherwise, it adds 1 to the year.
}
