% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{summarize_annual}
\alias{summarize_annual}
\title{Summarize annual data}
\usage{
summarize_annual(data, timeflag = "date_s")
}
\arguments{
\item{data}{data frame}

\item{timeflag}{character. Name of the time field, default is "date_s".}
}
\value{
data frame with annual summary
}
\description{
Summarize annual data
}
\details{
The function groups the data by TMSID2, TMSID, and year, and calculates the mean PM10 and PM25 values,
along with the number of days with data (ndays). It also replaces negative PM10 and PM25 values with NA.
}
