% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{summarize_daily}
\alias{summarize_daily}
\title{Summarize hourly data into daily}
\usage{
summarize_daily(data, timeflag = "datehour")
}
\arguments{
\item{data}{data frame}

\item{timeflag}{character}
}
\value{
data frame
}
\description{
Summarize hourly data into daily
}
\details{
The function groups the data by TMSID2, TMSID, and date, and calculates the mean PM10 and PM25 values.
It also flags days with insufficient data or excessive missing values.
}
