% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{tmb_rast}
\alias{tmb_rast}
\title{Convert \code{sdmTMB} result to terra}
\usage{
tmb_rast(tmb_res, log = FALSE)
}
\arguments{
\item{tmb_res}{\code{sdmTMB} result.}

\item{log}{logical(1). If \code{TRUE}, apply \code{exp} to the result.}
}
\value{
\code{terra} raster object.
}
\description{
Convert \code{sdmTMB} result to terra
}
